benteflork: Add basic hydra service.
This service doesn't have any build machines yet, but the database is
ready for use and the admin login works.

I'm using a Git submodule for Hydra here, because it's quite difficult
to import the Hydra source from a derivation. But maybe the machine is
able to bootstrap itself someday?

Signed-off-by: aszlig <aszlig@redmoonstudios.org>