Add custom lighttpd service.
The lighttpd module is somewhat different from the one in <nixos>, and
also is a bit more complicated, as it contains a way to declaratively
describe virtual hosts. I had that module in use for years at my
employer, but it never felt right to add it to <nixos> because I didn't
have time yet to add types and enough documentation.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>