Add a new pkgs tree for ESL's MongooseIM.
The package builds well with all dependencies, but doesn't run at the
moment, because we still need some patching and maybe even a constantly
running Erlang VM process which is used to do hot-reloads of future
updates.

But well, in the short term, just try to get it working even with the
need of VM restarts so we can do some testing and benchmarking.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>