lighttpd: Allow virtual host socket configuration.
This is a bit tricky, because the $SERVER["socket"] directive has to be
outside of the actual virtual host configuration, so we need to group
all socket directives together and put all the corresponding virtual
hosts within that directive.

One problem that is left is that when there is no socket attribute
present in the virtual host, the whole virtual host is discarded. For
global configuration we already have an option, but we might either
print an error or merge it with the global config in the future.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>