vhosts: Turn localCommands into a service.
In NixOS/nixpkgs@0e333688 mkOrder got removed and thus mkFooter is gone
as well. While it's not used within nixpkgs/nixos, we tried to correctly
order the localCommands and _append_ our subnets.

This now adds the subnet configuration as a systemd service, which
should not rely anymore on mkOrder and in addition it has the advantage
to be extendible by a stop action that shuts down those addresses
properly.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>