mongooseim: Add new service module.
Right now, this is only a basic service which creates a new Mnesia
database in the home directory of the "mongoose" user. It neither uses
hot code reloading nor is it particularily secure at the moment or
allows for flexible configuration.

But what works is running a basic XMPP server without too fancy
configuration.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>