hydra: Add kvm and nixos-test to all build slaves.
This only adds the features to the hydra machine, regardless on whether
the slave machines actually support it. But as we only have one slave
machine and it's using a NixOS default kernel, those features are indeed
supported.

Let's accurately detect this when we actually need to differentiate the
slave machines and their features.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>