tests/mongooseim: Refactor tests to use Escalus.
We can now run large parts of the test suite from ESL.

Unfortunately quite a lot here is unbelievably ugly and needs to
refactored even more. For example parts of the patching is done in the
mongooseimTests derivation, parts of them are done here at the tests.

Also, paths need to be cleaned up and there are a few hardcoded values,
like eg. the cookie.

However, most of the tests are already running successfully and this is
a big step forward instead of trying to write our own test suite for
ourselves.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>