Remove obsolete XMPP tests in Python.
We're using Escalus now, which is WAY better for testing XMPP servers
(primarily because its actually the goal of the application), so no need
to write crappy tests in Python.

Originally, I wanted to keep those tests as simple as possible, thus I
wen't for a quick and dirty Python script. But it probably wouldn't have
worked out very well in the long term, so I'm glad we made this decision
early.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>