ultron: Add a small Chromium update helper.
Probably this should better be implemented in Hydra, because it's a
hack. It basically makes a sparse checkout into the home directory of
the "nixpkgsupdate" user, calls update.sh and creates a commit of the
update (largely to avoid merge conflicts).

Hydra then picks up this directory and runs its builds. Of course, there
is a chance for race conditions if the evaluator picks up the directory
while update.sh is writing sources.nix.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>