mongooseim: Send startup notifications to systemd.
This is somewhat tricky in Erlang, as neither DBus nor Unix domain
sockets are available in the standard OTP library. So we add a new patch
which runs "systemd-notify" as an external program in order to inform
systemd not only about the fact that it has started but also on which
node it is running.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>