mongooseim: Add a small Erlang expression library.
This is going to be used for escaping values and/or in the future we
might want to even pretty-print values. But for now the library consists
of an attribute set with erl<TYPE> functions that do the necessary
escaping and one erlType function that can be used for mkOption's type
attribute.

The latter expects one argument, which is the type of the outermost
erlang expression. That way, we even have lightweight type checking for
Erlang expressions in Nix.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>