tests/mongooseim: Switch to new settings system.
Well, at least half of it, because the new settings system isn't quite
complete yet. Also, we needed to copy & paste the listener configuration
again, which is quite a bit redundant and shouldn't be necessary.

In the finished state it should be possible to override single options
of a particular listener, without the need to add the whole listener
again.

One particular issue here is that we can't easily unique together
multiple attributes in that option list, so we might want to switch to a
types.loaOf maybe?

Signed-off-by: aszlig <aszlig@redmoonstudios.org>