services/mongooseim: Add module configuration.
This currently neither has listed all the specific module options nor
does it use the correct defaults. For the latter, this is because we
currently cannot override specific module options, except if we copy &
paste most of the default configuration into the VM test case.

Also, there are quite a lot of modules missing from the original
Ejabberd server, so in the end we might want to port a few modules over
to MongooseIM.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>