Add XMPP configuration from old deployment.
This is based on the config of our current deployment, except for the
modules that are no longer available in MongooseIM. Another difference
is that the unused disco hosts are now gone and we're going to add them
whenever we actually have support for it.

However, there are still placeholders for IRC, ICQ and MSN which are to
be filled out later and are a requirement in order to actually deploy
our new systems.

The paths to the SSL certificates are also not existing and just dummy
values, which means that MongooseIM won't even start with this config,
so we're going to need a way to pass snake oil certificates for our
tests.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>