taalo: Switch from ext4 to btrfs.
Originally I've used ext4, because I thought it would be the most stable
and tested filesystem, but unfortunately it does have a problem with too
many directory links in /nix/store, leading to errors such as:

ext4_dx_add_entry:1551: Directory index full!

Of course, this can be mitigated by increasing the block size, but I've
ran fsck.ext4 with -D to recreate the directory index which ultimately
unlinked _all_ extents from /nix/store.

And yes, that was due to a bug in the version of e2fsprogs from the
rescue system.

So, the downside of btrfs however is that we can't make good assumptions
on the amount of free space available, but in the end I'd rather prefer
getting errors when actually hitting the physical disk limits than
needing to reformat with a different block size.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>