common: Deactivate packet filtering on all hosts.
This would introduce additional complexity for _no_ reason, because
every port which is listening to an external address needs to be
publicly accessible. So services are either not listening at all or only
locally and we don't provide shell access to users so there really is no
need for a packet filter.

Or in other words: You can't attack something that isn't there.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>