Fix handling of keys in test cases.
For real deployments, we still have deployment.keys, where we could set
owner/group permissions directly on the respective file. However, for
tests we generate snakeoil SSL keys where we really don't care whether
they're exposed.

So for tests, this means: Add all users of the services we're testing
and which are using keys to the keys group.

Also, as /run/keys became a ramfs in the meantime, we don't have POSIX
ACLs available anymore, so no 'setfacl'ing anymore ;-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>