modules: Add a new webspace service module.
Provides a simple way to map virtual hosts to user accounts and provide
a SFTP service for uploading content. The latter is using a seccomp BPF
sandboxed OpenSSH with a small patch which enables us to prefix user
names with "webspace_" and the client just uses the unprefixed version.

The webspace SFTP service will run on port 2222 by default, because it
is completely separate from the main SSHd. Essentially, this means that
we have to proxy the connection from the host system.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>