erlang-packages: Raw refactoring of repetitions.
Not yet to the degree I want it to be, but figuring that _all_ of those
packages are hosted on GitHub, we really can remove quite a lot of
repetetion and only use fetchFromGitHub for fetching sources.

Also note that a few of the SHA256 hashes are different, because we
previously either used fetchgit or fetchurl.

Package are named exactly as specified in src/*.app.src, except for
rebar_feature_runner, which has such a long name that my OCD with too
long lines kicked in and I'm now using rebarFR as the attribute name.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>