pkgs/mongooseim: Fix output dir of "ejabberd" app.
It seems that either reltool, robar or whichever is responsible for
generating the directory name doesn't seem to check the exit code of the
commend, so "$out/lib/ejabberd-sh: git: command not found\n2.1.8+mim-/"
ended up as a directory in the output store path. The output even should
be to stderr, so it seems that not only doesn't it care what's the error
code but also seems to merge stdin and stderr.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>