tests/mongooseim: Adapt tests to version 1.5.0.
Renames occurences of ejabberd with mongooseim and adds missing module
configurations that get added upstream since 1.2.2.

In version 1.5.0, we now have to provide a .spec file which lists the
test modules that are going to be included in the test run.

Also, we now have 5280 as the HTTP port for bosh, API and websocket
connections, so there is no need anymore to throw around with different
ports.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>