tests/mongooseim: Disable tests for mongooseimctl.
Unfortunately there are not only quite a lot of hardcoded paths in the
provided mongooseimctl script, it also depends on the "mongooseim"
script itself, which assumes that the system is using the SysV init
system.

We need to either need to heavily patch both scripts or write our own
implementation, before we could even try to re-enable these tests.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>