tests/mongooseim: Add missing test users.
These were included in test.config in the ejabberd_tests repository and
it makes sense to add them here as well, because we definitely want to
test even problematic stuff such as users using legacy clients.

Introduces nodeName2 and renames nodeName to nodeName1, because now the
tests need to be able to control both nodes.

Unfortunately, the latter currently seems to fail because we get a user
collision after the cluster set up. The latter however didn't really
happen, because we don't yet have support for mongooseimctl.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>