build-erlang: Install to $out/lib/erlang/lib.
Helps in resolving dependencies using code:lib_dir/1 (which is also used
by rebar to gather include paths), because the code server only accepts
to register names to paths containing the application name itself.

Another method would have been to patch rebar to rewrite the AST on
compilation of .erl files before the EPP runs (basically at the
tokenization phase) and rewrite every "-include_lib" attribute to
"-include" with a static path.

I initially thought this would be the cleaner solution, but after giving
it some thought, we already have $out/lib/erlang/lib for the Erlang
package in <nixpkgs> and it also keeps link farms clean (for example
when installing an Erlang lib with nix-env).

So it seems natural that every other library should follow that
structure as well, especially because patching rebar like mentioned
above would be error-prone as well, because code:lib_dir/1 would still
resolve to the wrong location (or give an error), so would need to patch
Erlang/OTP itself.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>