rebar-nix.patch: Improve handling of reltool deps.
It's not any nicer than before, because dependencies are still cramped
together in one single store path, but at least it's _working_, so it's
better than nothing.

The implementation is done by simply appending to the lib_dirs option in
reltool.config at runtime.

As of rebar/rebar#7 there should also be support for slim releases,
which in our case would be preferable not only for OTP but _all_ of the
dependencies, so we're certainly aiming to use this.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>