hydra-slave: Use separate /boot partition.
The new version of GRUB doesn't want to read mdraid level 0 properly.

We could have fixed this, but in the long run, it is more or less
guaranteed to be a constant source of errors. And event though it's just
a nixops destroy/deploy we really don't want to constantly reinstall the
build slaves on every occasion this will break again in the future.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>