services/mongooseim: Add option for Erlang cookie.
Includes quite a bit of refactoring, as we want to have the cookie
escaped correctly in order to make use of arbitrary characters rather
than just letters and numbers.

During tests however, we don't have the escaping function available, so
in the long run I guess it's a good idea to make our own systemd-aware
mongoosimctl utility, which correctly sets/escapes the cookie value.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>