xmppoke: Allow to configure the database backend.
Now, we're able to just override xmppoke with a databaseEngine attribute
to set whether we want SQLite or PostgrSQL. The main reason for this is
that we are more or less forced to use PostgrSQL because the upstream
SQL file for SQLite is too old and also seems to be unmaintained.

Also upgrades luadbi, because the release version had a bunch of
impurities regarding library/include paths.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>