tests/headcounter: Add subtests using xmppoke.
Right now, this is very ugly because we don't check whether the xmppoke
result is actually to our satisfaction. But at least now it works, even
though it is still slow as hell.

To improve the speed it's probably for the best to do one VM test per
domain.

We're also using PostgreSQL as the xmppoke database backend, where we
connect using the root user and trust authentication. It's just within a
VM test, so it doesn't matter whether it's really secure or not.

Another workaround right now is to list the domain names within the test
expression, so we may want to add a purpose flag to the vhost module.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>