tests/headcounter: Split into VM tests per vhost.
Sipping through all the virtual hosts withing one test run could take
quite a while, so let's parallelize the test runs and generate VM tests
for each vhost.

Especially if it comes to that endless loop in xmppoke at the moment, it
is a good idea to run it 4 times in parallel rather than sequentially
and waiting for all the timeouts one after another.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>