xmppoke: Use the same deps as in the squishy file.
First off, thanks a lot to @xnyhps for helping to debug this.

So, first of it turns out that we can't use the latest upstream Prosody
version, because server_select.lua shipped with xmppoke isn't compatible
with Prosody 0.9.7, so we now fetch it the same way as in the squishy
file, with one exception: We use util/hashes.so and util/encodings.so
from 0.9.7.

The reason for this is that the version used in the squishy file doesn't
define util.hashes.sha512 (well, to be more correct: it is defined but
"nil").

Next up, we need to shadow verse/{client,server}.lua and
verse/plugins/tls.lua from the xmppoke derivation. As xmppoke is first
in LUA's include path, the files from verse aren't even considered.

Finding out about verse/plugins/tls.lua took me some time, because the
cipher probe was stuck in an endless loop and it wasn't obvious (at
least to me) that this was the reason.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>