xmpp: Set OpenSSL ciphers.
This was quite a hard decision and I'm still not entirely sure about it,
but for now we want to have only ciphers providing forward secrecy, no
export or unauthenticated ciphers and no weak ciphers.

Unfortunately, it will prevent non-recent clients from connecting, so
maybe it makes sense to provide a small nag message for every client
which is connecting with a not quite secure enough cipher.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>