xmppoke: Use observatory frontend for reports.
Although the frontend is heavily patched by us to allow it to run
without a webserver and simply generate a static HTML output file.

We also pack together the full report into a tarball, so it doesn't
cause troubles if we want to put it through the shared folder in the VM
to the host system.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>