common: Split into common and common-machines.
In common.nix we're now defining things we need for *every* machine,
every VM test, basically every NixOS configuration in this repository.

Right now, this is just another import of module-list.nix and turning
off the damn netfilter, which is most of the duplication that has
infested the code base.

The common-machines.nix file is pretty much the same as the old
common.nix, except that it now imports the new common.nix as well and
thus turns off the packet filter.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>