Make custom packages available everywhere.
For NixOS modules, we just use nixpkgs.config.packageOverrides in
common.nix to make the headcounter namespace available and for
everything else, such as the tests, we now have a custom make-test.nix
expression, which not only passes our own packages but also the lib
attribute. Someday this might be useful if we want to have our custom
library as well.

However, one little caveat is that if we want to do real package
override in the parent path of our headcounter namespace we *will* have
an inconsistency between NixOS modules and non-module imports.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>