release: Fix build of manual.
Partially reverts changes from 3ae87ebf2d90746b997d5dcd0f6b4dadbda8958c.

For generating the manual, we still need to use module-list.nix directly
and avoid common.nix completely. In the end we only want Headcountr
specific options and nothing else.

So, this time, we're passing in the pkgs into the module args and don't
use packageOverrides at all. We also don't conflict with it because
we're not using common.nix for generating the manual.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>