modules: Add new service module for EPMD.
It's the Erlang Port Mapper Daemon and it's basically a small key/value
store where nodes register/unregister to/from it. The reason we're
making this its own service is that we want more control about when,
where and how EPMD is running and don't want to expose the EPMD port to
the outside world.

This module uses a patched version of the erlang derivation which
includes a combined diff of two commits that add support for systemd:

 * erlang/otp@59740d8fee66f3326b30ac25882b52ffada4bac2
 * erlang/otp@5db2345baa3bb9c835f91aafac3dde6e595debc7

As soon as OTP 17.4 hits the next NixOS stable release, we can simply
drop these patches.

One issue that is left for this module is to properly ensure that
to be registered services are already running for example while the VM
tests are trying to do a net_adm:ping/1.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>