modules: Add an imperative-containers module.
The module is to allow containers where the configuration.nix differs
from that of the whole headcounter deployment and is used to allow other
people to manage their containers.

Easily adding more such imperative containers and getting them up
quickly was the main motivation for writing this module.

Right now, quite a few stuff is hardcoded (like the motd) and thus not
suitable to be included in upstream NixOS and we need to have VM tests
in order to ensure that everything is working properly, but for now this
implementation works well enough and the configuration is going to be
overwritten on the first deployment of the container anyway.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>