hydra: Use signing for binary cache.
The narinfo files now should generate a signature on the fly using the
following public key:

headcounter.org:/7YANMvnQnyvcVB6rgFTdb8p5LG1OTXaO+21CaOSBzg=

We now also override the hydra-server systemd service to depend on
keys.target, so we can be sure that Hydra is only started as soon as all
keys are available.

Speaking of keys, of course the signing key is put into /run/keys and
the hydra user now is a member of the "keys" group, because only that
group (and of course root) has access to the /run/keys directory.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>