dns-server: Add current and new XMPP zones.
Those zones currently are only for transition/preparation to switch to
the new deployment. Note here, that the old XMPP server is "merkelig",
which has four IPv4 addresses:

 * 88.198.198.218 -> aszlig.net
 * 88.198.198.219 -> headcounter.org
 * 88.198.198.220 -> no-icq.org
 * 88.198.198.221 -> noicq.org

Unfortunately it only has *one* IPv6 address (2001:6f8:900:72a::2) using
a SixXS tunnel, which is going to be changed for our new deployment as
well but is kept here to ensure we don't introduce any additional false
positives when debugging.

Also note, that we now need two new GLUE records for ns1.headcounter.org
and ns2.headcounter.org.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>