dns-server: Always set serial to zero.
If I'm not mistaken *and* according to RFC1034, RFC1035 and RFC1982 the
serial value should only be relevant if we want to do zone transfers via
AXFR/IXFR.

But in our case, we deploy both nameservers simultaneously with the same
zone files and thus don't have any slaves at all.

Also, setting the serial to zero might cause problems with some
nameserver implementations. Citing from RFC1982:

Caution should also be exercised before causing the serial number to be
set to the value zero. While this value is not in any way special in
serial number arithmetic, or to the DNS SOA serial number, many DNS
implementations have incorrectly treated zero as a special case, with
special properties, and unusual behaviour may be expected if zero is
used as a DNS SOA serial number.

I dug into the source code of nsd and haven't found anything where it
treats a zero value specially, so I'm using it (because we can! YAY).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>