tests: Add a test to check if code reload works.
We want to make sure that the MongooseIM service doesn't *restart* on
changes, but only reloads the code + configuration.

This is implemented by a small test network consisting of a client with
escalus and a server with MongooseIM.

On the client, we run a small "testclient" Erlang server, which is
controlled using erl_call and it's more or less just a gen_server which
accepts atoms as messages and runs the tests in question.

After we've done some initialization (creating two test users and let
them communicate), we activate the new configuration on the server,
which disallows the use of C2S for all users.

Of course, if the server is *restarted*, the clients get disconnected,
so we check for disconnection shortly after the activation of the new
server configuration.

With the activation, we also *shouldn't* be able login anymore and we'll
check that plus the uptime of the server before and shortly after the
activation.

As we currently don't have hot code reloading this test obviously fails,
but the next step is to tackle issue #4 properly so we need that test.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>