mongooseim: Enable epmd service by default.
By default, Erlang tries to connect to an already running epmd instance,
but in order to prevent epmd to listen to all interfaces and to be a
proper systemd service, we enable it by default if the MongooseIM
service is enabled.

Of course, this means, we need to listen on all interfaces for the VM
tests as well, which would be the default behaviour on stock epmd.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>