pkgs/mongooseim: Remove nodetool and erl wrapper.
We don't use the nodetool script nor do we use the erl wrapper, so let's
not keep it around in the store path.

The reason why we don't use it that we have our own startup procedure
using systemd. If we'd need something like nodetool, we have to
implement it ourselves anyway.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>