modules: Add a new module for mongooseimctl.
We want to keep this module separate from the main MongooseIM module,
because the ctl command may be on a different machine while still
controlling an external node.

While this doesn't make much sense at the moment, it will have an impact
for the test suite where we have separated machines for the test runner
and the XMPP server.

Right now the only goof is that we have a hardcoded path to the ejabberd
application bundled with mongoosim (ejabberd-2.1.8+mim-1.5.1).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>