pkgs/mongooseim: Bail out on postBuild errors.
Picking up on 0a63c95, first there still was a "foo" left which I have
been using for testing. But more importantly, the postBuild phase
succeeded even though there was an error because of the mentioned "foo"
which I used for testing exactly that issue.

We now explicitly exit with status code 1 on any failing command to be
absolutely sure.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>