tests/mongooseim: Partially fix tests for ctl.
So far we had disabled the test for mongooseimctl (in 771ef00), but
having a NixOS module for the ctl command now, we can re-enable them.

Unfortunately the ejabberd_tests tries to find the ctl command based on
the cwd of the running MongooseIM node, which obviously won't work
because we're on a different machine.

The patch essentially searches PATH instead and because we have the
mongooseimctl module enabled, the client is able to find its way to
server1.

However, some tests are still failing and need to be fixed and/or
investigated why they do so.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>