pkgs/mongooseim: Update tests to latest master.
The current master version pretty much has quite a bunch of things
refactored. I was on the verge of patching a lot of stuff in that
repository but found that the latest master version has this already
refactored.

So I figured it would be a better idea to switch to master instead of
patching the hell out of the current release and patch everything again
after the next new upstream version has been released.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>