tests/mongooseim: Increase timeout of running ctl.
Currently running mongooseimctl requires a lot of time when running
inside of a VM, because /dev/random is pretty much starved the whole
time and thus generating a new UUID takes a ton of time.

So we need to fix mongooseimctl anyway, but it won't hurt to increase
the timeout anyway if we're on high load during test runs.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>