pkgs/mongooseim: Drop no-op sed call in postPatch.
Since esl/MongooseIM@56a5465, a call to "git describe" is no longer
issued and the version is fixed as the sed call was all about. That
means that we now can drop it safely without consequences because even
if the upstream project is going to reintroduce {vsn, {cmd, ...}}, we're
already checking for the main app directory on our side since 35cb996.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>