tests/code-reload: Fix testclient:register/3.
Escalus has been updated in 03a49a4, so we need to fix registration. The
issue here is that usually Escalus wants to have a path to a
configuration file. But we don't have that and we don't want that, so we
get around by mapping over all available users by ourselves.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>