mongooseim: Clean up service environment.
Let's move the configuration file to the serverArgsFile instead of
adding it via an environment variable, because this makes it easier to
change it via code reloads (using os:putenv/2 would work as well, but is
kinda hacky).

I'm not quite sure however about ROOTDIR, it doesn't seem to be directly
referenced in the MongooseIM code base so I'd guess that it's an old
relict. We can still bring it back when one of our VM tests fails.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>