mongooseim: Support configuration file reloading.
This currently makes the code-reload test happy, because we're only
supplying a new configuration file over there. But it won't solve #4
because whenever the store path of MongooseIM itself changes, we don't
have a way to upgrade the release yet.

Right now, we're creating a gcroot to the currently running store path,
so we have everything in place to check the differences between the
store path.

Turns out that we might want to write our own release handler, because
reltool only works with version numbers. Unfortunately, this won't work
on our side, because we'd need to increment a version number whenever we
have any changes.

In theory this could work if we'd create a Nix environment, so that we
have a generation number we can increment. But that will only work in
practice if we don't have any patches for mongooseim and all
applications we depend on.

To conclude: We currently exit with code 1 if the store paths differ
until we have either found a good way to deal with it using reltool or
we have our own implementation in place.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>