tests/code-reload: Actually test changing code.
So far, everything worked nicely because all we did was a small
configuration change and a check wether the new configuration applies to
the running server.

Unfortunately, if the server code stays the same and only the
configuration has changed, we leave the code alone and don't do a code
reload at all.

Now, we're really patching the code by simply substituting the "Pong"
reply message from mod_adhoc's ping command with a "Pang" and check
wether the result is returned accordingly.

Our VM test is now also separated into nice subtests so we get a nicer
HTML report.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>