build-erlang: Use sha256sums instead of versions.
Release utilities for Erlang all seem to work based on versions and so
do they also work if you want to do upgrades.

Fortunately it turns out that it doesn't matter whether the version is a
real version string or if it's a SHA256 hash, because there seems to be
no comparison whether a particular version is numerically higher or
lower than another one.

So, rewriting every version to use hashes obviously has the drawback
that whenever a particular application is going to reference that
version or do something with it (for example if we have preexisting
.appup files), this could get us in trouble.

In order to make sure this won't happen, we'd have to either check
whether this is the case and bail out if we can't fix it automatically
or we use a different scheme, for example using the hash _and_ the
version and patch application(3) to supply the correct version to the
application.

But until we have figured that out, we are fine if it comes to
MongooseIM, because it doesn't provide an .appup file by itself (yet).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>