pkgs: Add a new writeEscript function.
Makes it easy to either inline or reference an erlang file which is then
compiled into BEAM code and embedded into an escript file with a proper
shebang.

It's thus not only a shorthand for writeScript with the escript contents
directly but also executes faster because it doesn't need to be compiled
anymore.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>