build-erlang: Write appfile rewriter in Erlang.
Using sed to rewrite the app files is quite error-prone because we
cannot inspect the AST of Erlang. So using Erlang itself to rewrite
files is much less error-prone and also allows to implement other
versioning schemes more easily.

Unfortunately, file:consult/1 doesn't have a way to either specify the
encoding or directly strip off comments, so we have to use a first pass
which strips off all comments in binary mode and does the main rewrite
of the appfile in a second step.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>