build-erlang: Include version number in app files.
In 6da70f0, we implemented rewriting versions in app files with the
hashes from the Nix store path.

Now, we're appending the version again after the hash so we can easily
implement a way for applications to get the version number (for example
if the application provides an .appup file).

This is done in two different ways:

 * The main app file gets the version number specified in the version
   attribute from buildErlang.
 * Any application contained in the "apps" subdir gets the version
   number it originally provided and gets completely stripped if it's
   not a string (for example with {git, ...}).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>