mongooseim: Really build Erlang cookie locally.
The preferLocalBuild was actually set but unused, so it might be a good
idea to actually supply this to the derivation attributes.

Of course, this is to ensure that we won't pull the cookie file from our
public Hydra. The epmd process only listens on localhost, but let's be
more cautious and not even publically expose it in the first place.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>