xmpp: Allow the use of TLSv1.
It's quite ridiculous that there are quite a lot of operating systems
and clients out there which do not support TLSv1.2 or simply don't have
it enabled.

Switching to TLSv1.2 only in 1268dfd had the drawback that these clients
cannot connect anymore.

And yes, they're up to date, the problem is within the OS, for example
there are still quite a lot of devices out there which are on Android
version 4, which do not have TLSv1.2 enabled by default.

Forcing the users to upgrade their Android versions _would_ be an
option, but only if Android 5.x would be the latest version of their
device.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>