mongooseim: Update to latest upstream master.
This is in preparation for packaging the next upstream release, so we're
calling it 1.6.0pre instead of keeping it at 1.5.1 with a different git
revision.

For the upgrade we introduced a new dependency on exometer but dropped
the seestar dependency. The latter is only used for Apache Cassandra,
which we don't care about, so we can drop it.

Unfortunately, dropping these dependencies isn't so easy, so we need to
introduce a strip-unneeded-deps.patch that's getting rid of all
references to mysql, pgsql, seestar and riakc.

The patch is also supported by a new prePatch phase, which deletes all
files and directories for supporting these databases.

Instead of using the postBuild phase (see 0a63c95) to get rid of the
nodetool and erl wrappers, we're doing so now in the reltool.patch,
among with all other cruft that is created by the upstream reltool.conf
like for example mongooseimctl (we have our own implementation).

And last but not least, let's drop mod_metrics from the NixOS service,
because that module has been dropped in 1.5.1 already but it's now
causing an error on startup.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>