xmpp: Support DHE-*-AES256-SHA ciphers.
I'm very much reluctant to do this, but unfortunately we can't get away
from this because some clients like Gajim on Windows or Adium on OSX are
still using OpenSSL version 0.9.8.

Now, basically OpenSSL version 0.9.8 would support better options such
as ECDH-ECDSA-AES256-SHA, but unfortunately only for SSLv3 and not TLS,
see this mail for an explanation:

http://marc.info/?l=openssl-users&m=132065751332665

Fixes issue #18.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>