mongoosiem/modules: Filter out _module attribute.
It was originally called __internal and was introduced in
NixOS/nixpkgs@1d62ad4, later in NixOS/nipkgs@05e8a48 it has been renamed
to _module.

This attribute is exposed to all modules and at the toplevel of
submodules, so we need to filter it out for generating the list of
enabled modules.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>