hydra: Update to latest upstream master.
So far I was running with a local clone of the source code with patches,
but since NixOS/hydra#211 is now merged, we can use the upstream version
without changes.

However, this also comes with a merge of the build-ng branch (in
NixOS/hydra@b0c8eec), which introduces quite a lot of changes on the
administrative side.

First of all, the consecutive Hydra services now use diffent users, so
we need to give them privileges to keys.

Second, we no longer need to set PostgreSQL to use peer authentication,
because that's now the default of the upstream module.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>