hydra: Don't force nix.maxJobs to 0.
This had an affect at the time where Hydra was using the distributed
build feature of Nix itself. The problem was that the distributed build
feature was scheduling jobs on the Hydra master, which could be quite
poisionous in terms of machine load.

Since the merge of Hydra's build-ng branch, this is no longer an issue
and we can safely drop it.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>