tests/code-reload: Properly reference "pkgs" attr.
Using pkgs propagated from make-test.nix works from within the test
machine configurations but unfortunately won't apply overrides, because
we have them in common.nix of the corresponding machine configuration.

So turning the configurations into a function with shadowed pkgs fixes
this problems and the overrides should be applied properly.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>