ssl: Temporarily switch to StartSSL certificates.
In the long run I'm going to replace them by letsencrypt certificates,
but because the certificates were about to expire I had use a CA to use
in the mean time.

Note that we now have certificates with jabber.* and an alias to the
main FQDN, although the jabber.* subdomains are currently unused.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>